Restructure shared
