+++ src/client/app/shared/asy-transclude.directive.ts
@@ -1,4 +1,9 @@
-import { Directive, Input, ViewContainerRef, TemplateRef, OnChanges, SimpleChange, SimpleChanges, EmbeddedViewRef } from '@angular/core';
+import {
+	Directive, Input, ViewContainerRef,
+	TemplateRef, OnChanges, SimpleChange,
+	SimpleChanges, EmbeddedViewRef
+} from '@angular/core';
+
 @Directive({
 	selector: '[asy-transclude]'
 })
