+++ src/client/app/shared/camel-to-human.pipe.ts
@@ -1,4 +1,5 @@
 import {Pipe, PipeTransform} from '@angular/core';
+
 import {StringUtils} from './string-utils.service';
 
 @Pipe({name: 'camelToHuman'})
