+++ src/client/app/shared/dropdown/index.ts
@@ -2,4 +2,6 @@ export * from './asy-dropdown-item-wrapper.component';
 export * from './asy-dropdown-item.component';
 export * from './asy-dropdown.component';
 
+export * from './asy-dropdown-type.class';
+
 export * from './asy-dropdown.service';
