+++ src/client/app/shared/observable-utils.class.ts
@@ -1,4 +1,5 @@
 import { Observable } from 'rxjs';
+
 import { ObservableResult } from './observable-result.class';
 
 export class ObservableUtils {
